[MOD] : transfer 필드 타입 변경

+++ app/src/main/java/com/sopt/kakaobank/presentation/transfer/Transfer.kt
@@ -5,6 +5,6 @@ data class Transfer(
     val profile: String,
     val accountName: String,
     val bankName: String,
-    val accountNumber: Int,
+    val accountNumber: Long,
     var bookmark: Boolean,
 )
